Make PartialGerberCode trait non-public
